webapp/requirements.txt
flask>=3.0.0
pillow>=10.0.0
torch>=2.2.0
torchvision>=0.17.0
